Update materials for resurfacing
