+++ primer-on-python-decorators/README.md
@@ -1,17 +1,11 @@
 # Primer on Python Decorators
 
-Code examples from the [Primer on Python Decorators](https://realpython.com/primer-on-python-decorators/) tutorial on [Real Python](https://realpython.com/).
+Here you can find code examples from the [Primer on Python Decorators](https://realpython.com/primer-on-python-decorators/) tutorial on [Real Python](https://realpython.com/).
 
 ## Decorators
 
-As noted in the article, most decorators are stored in the file [`decorators.py`](https://github.com/realpython/materials/blob/master/primer-on-python-decorators/decorators.py). The exceptions are those decorators that depend on third party packages ([Flask](http://flask.pocoo.org/) and [Pint](https://pint.readthedocs.io/)), which are available in [`decorators_flask.py`](https://github.com/realpython/materials/blob/master/primer-on-python-decorators/decorators_flask.py) and [`decorators_unit.py`](https://github.com/realpython/materials/blob/master/primer-on-python-decorators/decorators_unit.py), respectively.
+You can find most decorators in the file `decorators.py`. The exceptions are those decorators that depend on the third-party packages ([Flask](http://flask.pocoo.org/) and [Pint](https://pint.readthedocs.io/)). These decorators are available in `secret_app.py`, `validate_input.py`, and `units.py`.
 
 ## Examples
 
-Most of the code examples from the article are available in the [`examples.py`](https://github.com/realpython/materials/blob/master/primer-on-python-decorators/examples.py) file.
-
-## Cheat Sheet
-
-We’ve put together a short & sweet Python decorators cheat sheet for you that summarizes the techniques explained in this tutorial:
-
-[Get the decorators cheat sheet »](https://realpython.com/optins/view/decorators-cheatsheet/)
+You can find many code examples from the tutorial in separate Python files.
