+++ primer-on-python-decorators/class_decorators.py
@@ -1,0 +1,22 @@
+from decorators import debug, timer
+
+
+class TimeWaster:
+    @debug
+    def __init__(self, max_num):
+        self.max_num = max_num
+
+    @timer
+    def waste_time(self, num_times):
+        for _ in range(num_times):
+            sum([number**2 for number in range(self.max_num)])
+
+
+@timer
+class TimeWaster2:
+    def __init__(self, max_num):
+        self.max_num = max_num
+
+    def waste_time(self, num_times):
+        for _ in range(num_times):
+            sum([i**2 for i in range(self.max_num)])
