+++ primer-on-python-decorators/inner_functions.py
@@ -1,0 +1,33 @@
+def parent():
+    print("Printing from parent()")
+
+    def first_child():
+        print("Printing from first_child()")
+
+    def second_child():
+        print("Printing from second_child()")
+
+    second_child()
+    first_child()
+
+
+def get_child(num):
+    def first_child():
+        return "Hi, I'm Elias"
+
+    def second_child():
+        return "Call me Ester"
+
+    if num == 1:
+        return first_child
+    else:
+        return second_child
+
+
+if __name__ == "__main__":
+    parent()
+
+    first = get_child(1)
+    second = get_child(2)
+    print(first())
+    print(second())
